serviceapp: use only seekTo functions of external players
exteplayer3 relative seeking is imprecise, since it's using last written
pts to decoder as a starting point for seek, instead of pts of shown frame.

This should be fixed on exteplayer3 side, but we really don't need to
use it, since we could compute next position and just use seekTo
functions.

Note there is also added end protection, since if on vuplus we seek
out of bounds pts is not updated, but vuplus drivers continue to
respond to GET_PTS ioctl calls by increasing pts, which breaks exteplayer3
EOS detection