serviceapp: enable auto turning on subtitles
InfoBarSubtitlesSupport enables subtitle stream which it gets from
getCachedSubtitle.
getCachedSubtitle returns first subtitle stream if auto turning on
config is enabled and there are some subtitles available.
getCachedSubtitle is queried on evUpdatedInfo event, which we sent
from eServiceApp one second after player process has started.
One second is there because on player's start it's possible that not all
subtitles streams are discovered, maybe we need to make it even bigger.