improve auto-selection and exposure of external subtitles
- all external subtitles are available for selection withing video
directory, it will also look for subtitles in "Subs" directory
in video directory.

- external subtitles don't have language information so instead
of setting it to "unk" set it to filename which will provide little
more info to user.

- auto selection when not preferring embedded subtitles (default):

1. select external subtitle if subtitle has the same name as video
without extension

2. select external subtitle if there is only one listed in video and
"Subs" directory - this is done because we really don't know which one
to choose if there are more of them. We could not choose any subtitle
but it's convenient to not need to always rename subtitle so it corresponds to
video filename, it could possibly be invalid subtitle, but its still
good trade for occasional mistake.

3. select first embedded subtitle