m3u8: store used User-Agent header in headers
Some servers don't like when different user-agents are used
in successive requests, i.e. they respond with 403 forbidden status.

This happens when user doesn't provide user-agent and default one is used
to retrieve master playlist. Then url to variant is passed to player
which uses it's own user-agent string, since none was set.

To fix this we store user-agent header in headers when retrieving
variants from master playlist, so player will pick it up and use it.